rich message factory code refactoring
