+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/conversation/richmessaging/adapters/AlRichMessageAdapterFactory.java
@@ -4,6 +4,7 @@ import android.content.Context;
 
 import com.applozic.mobicomkit.api.conversation.Message;
 import com.applozic.mobicomkit.uiwidgets.AlCustomizationSettings;
+import com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.AlRichMessageFactory;
 import com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.callbacks.ALRichMessageListener;
 import com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.models.ALRichMessageModel;
 
@@ -18,14 +19,6 @@ import java.util.Map;
  * created on: 23 Nov, 2019
  */
 public class AlRichMessageAdapterFactory {
-    //constants for specifying which instance to instantiate
-    public final static int CARD_RICH_MESSAGE = 10;
-    public final static int IMAGE_RICH_MESSAGE = 9;
-    public final static int FAQ_RICH_MESSAGE = 8;
-    public final static int LIST_RICH_MESSAGE = 7;
-    public final static int BUTTON_RICH_MESSAGE = 3;
-    public final static int REPLY_RICH_MESSAGE = 6;
-
     //factory class is a singleton
     private AlRichMessageAdapterFactory() {
     }
@@ -40,11 +33,11 @@ public class AlRichMessageAdapterFactory {
     }
 
     public ALRichMessageAdapter getRMAdapter(Context context, ALRichMessageModel model, ALRichMessageListener listener, Message message) {
-        if (model.getTemplateId() == AlRichMessageAdapterFactory.CARD_RICH_MESSAGE)
+        if (model.getTemplateId() == AlRichMessageFactory.CARD_RICH_MESSAGE) {
             return new AlCardRMAdapter(context, model, listener, message);
-        else if (model.getTemplateId() == AlRichMessageAdapterFactory.BUTTON_RICH_MESSAGE || model.getTemplateId() == AlRichMessageAdapterFactory.REPLY_RICH_MESSAGE)
+        } else if (model.getTemplateId() == AlRichMessageFactory.BUTTON_RICH_MESSAGE || model.getTemplateId() == AlRichMessageFactory.REPLY_RICH_MESSAGE || model.getTemplateId() == AlRichMessageFactory.MIXED_BUTTON_RICH_MESSAGE) {
             return new AlButtonRMAdapter(context, model, listener, message);
-        else return null;
+        } else return null;
     }
 
     public ALRichMessageAdapter getImageRMAdapter(Context context, ALRichMessageModel model, ALRichMessageListener listener, Message message, AlCustomizationSettings alCustomizationSettings) {
